64 bit fixes. an int is not big enough on 64 bit platforms to store addresses. Use longs instead which will be correct on both 32bit and 64bit systems.

+++ elfchk/dwarf.h
@@ -35,8 +35,8 @@ typedef struct CIEFrame {
 typedef struct FDEFrameHeader {
 	Elf32_Word	length;
 	Elf32_Word	CIE_pointer;
-	int		*initial_location;
-	int		address_range;
+	long		*initial_location;
+	long		address_range;
 	} FDEFrameHeader;
 
 typedef struct EHFRMHDRImage {
